Modified to support 10 s refresh rate

+++ gen_data.sh
@@ -6,12 +6,19 @@ ENDPATH=/glade/work/csgteam/qs_cache
 mkdir -p $ENDPATH
 cd $TMPPATH
 
+# Don't run if already running
+if [[ -f qscache-active ]]; then
+    exit
+fi
+
+touch qscache-active
+
 # Get data from PBS (must run as csgteam)
 sudo -u csgteam /opt/pbs/bin/qstat -a -1 -n -s -w -x | sed '1,5d' > newlist-wide.dat &
 sudo -u csgteam /opt/pbs/bin/qstat -1 -n -s -x | sed '1,5d' > newlist-info.dat &
 sudo -u csgteam /opt/pbs/bin/qstat -x | sed '1,5d' > newlist-default.dat &
 
-time wait
+wait
 
 # Poor-man's sync
 mv newlist-wide.dat commlist-wide-nodes.dat
@@ -22,7 +29,7 @@ mv newlist-default.dat joblist-default.dat
 grep -v '^ ' commlist-wide-nodes.dat > joblist-wide-nodes.dat &
 grep -v '^ ' commlist-info-nodes.dat > joblist-info-nodes.dat &
 
-time wait
+wait
 
 # Get versions without nodelist
 sed -r 's/([0-9].*) [-,r].*/\1/' commlist-wide-nodes.dat > commlist-wide.dat &
@@ -30,7 +37,8 @@ sed -r 's/([0-9].*) [-,r].*/\1/' commlist-info-nodes.dat > commlist-info.dat &
 sed -r 's/([0-9].*) [-,r].*/\1/' joblist-wide-nodes.dat > joblist-wide.dat &
 sed -r 's/([0-9].*) [-,r].*/\1/' joblist-info-nodes.dat > joblist-info.dat &
 
-time wait
+wait
 
 # Move files to final storage
 mv *.dat $ENDPATH
+rm qscache-active
